Add basic functionality
The timing package can be used to keep track of timing information
within a function.  You simply call timing.New() to create a new timing
instance, then call Time("foo") to start a timing with the label "foo"
and TimeEnd("foo") to stop the timer.  Calling String() returns a string
containing all of the finalized labels and their associated duration.